# docs/icd/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The HTTP interface of one Sentinel node: hub, edge or standalone. Every node serves its own console from its own data (ADR-009), so nothing here depends on the link to another node. Times are ISO 8601 UTC and units are field-name suffixes (`_m`, `_m_s`, `_km`, `_s`). A read-only node answers every write with 403. The acting operator is named by the `X-Sentinel-Operator` header, set by the front proxy. Interface control documents: docs/icd/README.md.",
    "title": "Sentinel",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/demo/link": {
      "post": {
        "description": "Demonstration and harness only: drives the link emulator. 403 unless demo controls\nare on and the caller is on localhost; 503 without an emulator; 422 on an unknown\npreset.",
        "operationId": "demo_link_api_demo_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "preset": {
                    "enum": [
                      "CONNECTED",
                      "DEGRADED",
                      "DENIED",
                      "LIMITED"
                    ]
                  }
                },
                "required": [
                  "preset"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Demo Link Api Demo Link Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Apply a link-emulation preset (demo only)",
        "tags": [
          "link and sync"
        ]
      }
    },
    "/api/events": {
      "get": {
        "description": "`scope`: `active` (TCA in the future, soonest maneuver commit point first),\n`past` or `all` (newest TCA first); 400 otherwise. An edge also lists events it\nknows only from the hub's summaries, marked `verification: HUB_ASSERTED` until\ntheir CDMs arrive and are re-assessed locally.",
        "operationId": "events_api_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "active",
              "title": "Scope",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Events Api Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List conjunction events",
        "tags": [
          "conjunctions"
        ]
      }
    },
    "/api/events/{event_id}": {
      "get": {
        "description": "The event summary, every CDM received for it with its own assessment and\nadmission warnings, and the engine version and policy that produced them. 404 when\nthe event is unknown.",
        "operationId": "event_api_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Event Api Events  Event Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One event: summary and CDM history",
        "tags": [
          "conjunctions"
        ]
      }
    },
    "/api/events/{event_id}/annotation": {
      "post": {
        "description": "Writes a multi-value register. Concurrent writes on different nodes are all kept\nand shown as a conflict, never silently resolved. 422 on an unknown field or triage\nstatus; 403 on a read-only node.",
        "operationId": "annotation_api_events__event_id__annotation_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "field": {
                    "enum": [
                      "triage_status",
                      "assignee",
                      "note"
                    ]
                  },
                  "value": {
                    "description": "for triage_status, one of NEW, WATCH, MANEUVER_PLANNING, NO_ACTION, CLOSED"
                  }
                },
                "required": [
                  "field",
                  "value"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Annotation Api Events  Event Id  Annotation Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set an annotation",
        "tags": [
          "operator data"
        ]
      }
    },
    "/api/events/{event_id}/decision": {
      "post": {
        "description": "Appends a signed DECISION to the log, bound to the event's current CDM. 422 on a\ndecision outside the list; 403 on a read-only node.",
        "operationId": "decision_api_events__event_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "decision": {
                    "enum": [
                      "MANEUVER",
                      "NO_MANEUVER",
                      "MONITOR",
                      "REQUEST_TASKING"
                    ]
                  },
                  "rationale": {
                    "description": "truncated to 2000 characters",
                    "type": "string"
                  }
                },
                "required": [
                  "decision"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Decision Api Events  Event Id  Decision Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record a decision",
        "tags": [
          "operator data"
        ]
      }
    },
    "/api/events/{event_id}/dilution-curve": {
      "get": {
        "description": "Pc as the covariance is scaled by k (log10 grid), with the maximizing scale k*.\n`diluted` is true when k* < 1: the Pc would rise if the data were better.\n`available: false` when the curve cannot be computed.",
        "operationId": "dilution_curve_api_events__event_id__dilution_curve_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dilution Curve Api Events  Event Id  Dilution Curve Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pc against covariance scale",
        "tags": [
          "conjunctions"
        ]
      }
    },
    "/api/events/{event_id}/encounter": {
      "get": {
        "description": "Miss vector, projected covariance ellipse and hard-body radius in the encounter\nplane, for display. `available: false` with a reason when there is no covariance or\nno hard-body radius.",
        "operationId": "encounter_api_events__event_id__encounter_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Encounter Api Events  Event Id  Encounter Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Encounter-plane geometry",
        "tags": [
          "conjunctions"
        ]
      }
    },
...
  },
...
}